Feat/refactor clustering (#94)
