--- kura/cluster/models/__init__.py
@@ -1,3 +1,0 @@
-from .kmeans import KmeansClusteringModel
-
-__all__ = ["KmeansClusteringModel"]
