--- kura/cluster/prompts.py
@@ -1,34 +1,0 @@
-DEFAULT_CLUSTER_PROMPT = """
-You are tasked with summarizing a group of related statements into a short, precise, and accurate description and name. Your goal is to create a concise summary that captures the essence of these statements and distinguishes them from other similar groups of statements.
-
-Summarize all the statements into a clear, precise, two-sentence description in the past tense. Your summary should be specific to this group and distinguish it from the contrastive answers of the other groups.
-
-After creating the summary, generate a short name for the group of statements. This name should be at most ten words long (perhaps less) and be specific but also reflective of most of the statements (rather than reflecting only one or two).
-
-The name should distinguish this group from the contrastive examples. For instance, "Write fantasy sexual roleplay with octopi and monsters", "Generate blog spam for gambling websites", or "Assist with high school math homework" would be better and more actionable than general terms like "Write erotic content" or "Help with homework". Be as descriptive as possible and assume neither good nor bad faith. Do not hesitate to identify and describe socially harmful or sensitive topics specifically; specificity is necessary for monitoring.
-
-Present your output in the following format:
-<summary> [Insert your two-sentence summary here] </summary>
-<name> [Insert your generated short name here] </name>
-
-The names you propose must follow these requirements:
-- The cluster name should be a sentence in the imperative that captures the user's request. For example, 'Brainstorm ideas for a birthday party' or 'Help me find a new job.'
-- Create names that are specific enough to be meaningful, but not so specific that they can't meaningfully represent many different statements.
-- Avoid overly general or vague terms, and do not hesitate to describe socially harmful or sensitive topics (in fact, clusters that clearly describe harmful behavior are slightly preferred); specificity is necessary for observability and enforcement.
-- Ensure that the cluster name is distinct from the contrastive examples.
-- Use clear, concise, and descriptive language for the cluster name.
-
-Below are the related statements:
-<positive_examples>
-{% for item in positive_examples %}{{ item }}
-{% endfor %}
-</positive_examples>
-
-For context, here are statements from nearby groups that are NOT part of the group you're summarizing:
-<contrastive_examples>
-{% for item in contrastive_examples %}{{ item }}
-{% endfor %}
-</contrastive_examples>
-
-Do not elaborate beyond what you say in the tags. Remember to analyze both the statements and the contrastive statements carefully to ensure your summary and name accurately represent the specific group while distinguishing it from others. The cluster name should be a sentence in the imperative that captures the user's request. For example, 'Brainstorm ideas for a birthday party' or 'Help me find a new job.'
-"""
