--- kura/embedding/__init__.py
@@ -1,4 +1,0 @@
-from .embedding import embed_summaries
-from .utils import batch_texts
-
-__all__ = ["embed_summaries", "batch_texts"]
