--- kura/embedding/models/__init__.py
@@ -1,4 +1,0 @@
-from .openai_embedding_model import OpenAIEmbeddingModel
-from .sentence_transformers_embedding_model import SentenceTransformerEmbeddingModel
-
-__all__ = ["OpenAIEmbeddingModel", "SentenceTransformerEmbeddingModel"]
